We androidX now
Summary:
Migrating to android X for next major release on 5th May.
Reason being is a few partners are asking for it, and it is nice to use shiny new things.

At the same time updating travis to use android x sdk.

I plan to rebase after our release on 20th April, but want to hear your opinions and concerns now.

Reviewed By: jingping2015

Differential Revision: D20074358

fbshipit-source-id: 7f88e641eaf1401ba9ccf174dcf1543d16aa8656